add empty main screen
